New exp calulation
exp = 25 + current health + 1 for each 20 exp you have

+++ Assets/NewProject/scripts/Player/PlayerBase.cs
@@ -45,7 +45,7 @@ public class PlayerBase : MonoBehaviour
     public float actionPoints;
     public float maxActionPoints;
     public float range;
-    public int exp = 0;
+    public float exp = 0;
     private OG_MovementByMouse checkMovement;
     public PlayerActionManager turnsDone;
